Gates

+++ app/Http/Controllers/Admin/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 class UsersController extends Controller
@@ -64,6 +65,9 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('edit-users')){
+            return redirect(route('admin.users.index'));
+        }
         $roles = Role::all();
 
         return view('admin.users.edit')->with([
@@ -94,6 +98,9 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('delete-users')){
+            return redirect(route('admin.users.index'));
+        }
         $user->roles()->detach();
         $user->delete();
